Mapeada relación entre entidades TipoEquipo, Equipo y Miembros. Cambiados ids de tablas de Integer a Long, añadidas secuencias. Renombrada vista y redirigida entrada menú de listadoEquipos a tiposEquipo ya que es la vista con registros tabla TipoEquipo. Eliminada opción de equipos especiales. Añadida busqueda de equipos por nombre de un miembro (componente) y por tipo de equipo. Adaptados métodos, vistas y script sql a la nueva situación.

+++ src/main/java/es/mira/progesin/persistence/repositories/ITipoEquiposRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import es.mira.progesin.persistence.entities.TipoEquipo;
 
-public interface ITipoEquiposRepository extends CrudRepository<TipoEquipo, Integer> {
+public interface ITipoEquiposRepository extends CrudRepository<TipoEquipo, Long> {
 
 }
